SDK supports both singleton and non-singleton usage
