--- inlong-audit/audit-sdk/src/main/java/org/apache/inlong/audit/DefaultISocketAddressListLoader.java
@@ -1,63 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.inlong.audit;
-
-import org.apache.commons.lang3.StringUtils;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-/**
- * DefaultISocketAddressListLoader
- */
-public class DefaultISocketAddressListLoader implements SocketAddressListLoader {
-
-    public static final String KEY_AUDIT_PROXYS = "audit.proxys";
-
-    private Map<String, String> commonProperties;
-
-    @Override
-    public void setCommonProperties(Map<String, String> commonProperties) {
-        this.commonProperties = commonProperties;
-    }
-
-    /**
-     * loadSocketAddressList
-     * @return
-     */
-    @Override
-    public List<String> loadSocketAddressList() {
-        List<String> ipPortList = new ArrayList<>();
-        String strAuditProxys = commonProperties.get(KEY_AUDIT_PROXYS);
-        if (strAuditProxys == null) {
-            return ipPortList;
-        }
-        String[] ipPorts = strAuditProxys.split("\\s+");
-        for (String tmpIPPort : ipPorts) {
-            if (StringUtils.isBlank(tmpIPPort)) {
-                continue;
-            }
-            ipPortList.add(tmpIPPort.trim());
-        }
-        Collections.sort(ipPortList);
-        return ipPortList;
-    }
-
-}
